Move cluster setup into AbstractPrometheusRestIT
The ElasticsearchCluster @ClassRule and getTestRestCluster() were
identical across all 6 IT classes. Moving them to the abstract base
removes ~114 lines of duplication, following the same pattern used by
AbstractOTLPIndexingRestIT.
