+++ x-pack/plugin/prometheus/src/javaRestTest/java/org/elasticsearch/xpack/prometheus/PrometheusLabelValuesRestIT.java
@@ -11,9 +11,6 @@ import org.apache.http.client.utils.URIBuilder;
 import org.elasticsearch.client.Request;
 import org.elasticsearch.client.Response;
 import org.elasticsearch.client.ResponseException;
-import org.elasticsearch.test.cluster.ElasticsearchCluster;
-import org.elasticsearch.test.cluster.local.distribution.DistributionType;
-import org.junit.ClassRule;
 
 import java.io.IOException;
 import java.util.List;
@@ -33,22 +30,6 @@ import static org.hamcrest.Matchers.notNullValue;
  * Detailed plan-building and response-parsing logic is covered by unit tests.
  */
 public class PrometheusLabelValuesRestIT extends AbstractPrometheusRestIT {
-
-    @ClassRule
-    public static ElasticsearchCluster cluster = ElasticsearchCluster.local()
-        .distribution(DistributionType.DEFAULT)
-        .user(USER, PASS, "superuser", false)
-        .setting("xpack.security.enabled", "true")
-        .setting("xpack.security.autoconfiguration.enabled", "false")
-        .setting("xpack.license.self_generated.type", "trial")
-        .setting("xpack.ml.enabled", "false")
-        .setting("xpack.watcher.enabled", "false")
-        .build();
-
-    @Override
-    protected String getTestRestCluster() {
-        return cluster.getHttpAddresses();
-    }
 
     public void testInvalidSelectorSyntaxReturnsBadRequest() throws Exception {
         Request request = labelValuesRequest("job", "{not valid!!!}");
